refactor(awq): unify sidecar loader into one pub fn (#268 items 2 + 3)
Three near-identical copies of the AWQ sidecar load path existed (hfq.rs
closure, qwen35.rs::load_awq_scale_for, dflash.rs inlined block); the
dflash.rs comment even self-documented it as "structurally a third copy".
Consolidate into `pub fn hipfire_runtime::hfq::load_awq_scale` and call
from all three sites.

Internally the helper routes through `tensor_data_vec`, which on Unix
already does pread + fadvise_dontneed (avoids page cache buildup on
unified-memory APUs) and on non-Unix falls back to mmap. Same byte path
into GPU, slightly safer load behavior at the two previously mmap-only
sites (hfq.rs, dflash.rs).

Also fixes the cast direction at the former hfq.rs:518 — the old `!=
k as u32` shape check is now an `as usize` on the u32 side (item 2 of
#268). Harmless in practice (max model K ~50000) but the new helper is
correct by construction.

Closes #268 items 2 + 3. Items 1 (defensive scale clamp), 4 (router
whitelist bench), and 5 (reciprocal precompute) remain open.
