+++ crates/hipfire-arch-qwen35/src/qwen35.rs
@@ -5,7 +5,7 @@
 //! Qwen3.5 model: hybrid DeltaNet (linear attention) + standard attention.
 //! Feature-gated behind `deltanet`.
 
-use hipfire_runtime::hfq::HfqFile;
+use hipfire_runtime::hfq::{load_awq_scale, HfqFile};
 use hipfire_runtime::llama::{self, f16_to_f32, EmbeddingFormat, WeightTensor, weight_gemv,
                               weight_gemv_prerotated, fused_rmsnorm_rotate_for_mq,
                               fused_rmsnorm_rotate_mq_batched_for,
@@ -842,52 +842,6 @@ fn load_weight_tensor_raw(gpu: &Gpu, quant_type: u8, data: &[u8], m: usize, k: u
     }
 }
 
-/// Phase A Stage A — AWQ sidecar loader for the Qwen3.5 forward path.
-///
-/// The .hfq quantizer emits `<weight>.awq_scale.weight` (1D F16, length K)
-/// alongside MQ4G256 weights that were AWQ pre-scaled. The dispatcher in
-/// `fused_rmsnorm_rotate_for_mq` / `fused_rmsnorm_rotate_mq_batched_for`
-/// looks at `WeightTensor.awq_scale.is_some()` to pick the AWQ-aware
-/// kernel variant. WITHOUT this loader populating the field, every MQ4
-/// weight ends up with `awq_scale: None`, the dispatcher falls through
-/// to the non-AWQ kernel, and the math `(W·s) · (x/s) = W·x` breaks
-/// because the runtime never divides by `s` — observed KLD blowup
-/// 0.6721 → 13.4893 on 0.8B Qwen3.5 before this landed.
-///
-/// Lookup pattern matches `hipfire_runtime::hfq::load_awq_scale`:
-/// strip trailing `.weight`, append `.awq_scale.weight`. Try both the
-/// `model.language_model.`-prefixed name and the bare name (the qwen35
-/// crate uses prefixed names; older sidecars or tests may use either).
-fn load_awq_scale_for(hfq: &HfqFile, gpu: &Gpu, name: &str, k: usize) -> Option<GpuTensor> {
-    let sidecar_name = match name.strip_suffix(".weight") {
-        Some(stem) => format!("{stem}.awq_scale.weight"),
-        None => format!("{name}.awq_scale.weight"),
-    };
-    let (sc_info, sc_data) = hfq.tensor_data_pread(&sidecar_name)?;
-    // Must be 1D F16, length K. quant_type 1 = F16.
-    if sc_info.quant_type != 1 {
-        eprintln!(
-            "warning: AWQ sidecar {sidecar_name} has quant_type={} (expected 1=F16); skipping",
-            sc_info.quant_type
-        );
-        return None;
-    }
-    if sc_info.shape.len() != 1 || sc_info.shape[0] as usize != k {
-        eprintln!(
-            "warning: AWQ sidecar {sidecar_name} shape mismatch ({:?} vs expected [{}]); skipping",
-            sc_info.shape, k
-        );
-        return None;
-    }
-    // F16 → F32 on host so the kernel takes a plain `const float*`.
-    let f32_data: Vec<f32> = sc_data
-        .chunks_exact(2)
-        .map(|c| f16_to_f32(u16::from_le_bytes([c[0], c[1]])))
-        .collect();
-    let f32_bytes: Vec<u8> = f32_data.iter().flat_map(|&v| v.to_le_bytes()).collect();
-    gpu.upload_raw(&f32_bytes, &[f32_bytes.len()]).ok()
-}
-
 /// TODO(transformer-extraction): cross-arch duplicate of
 /// `hipfire-arch-qwen2::qwen2::load_weight_tensor` — same name-lookup +
 /// pread + AWQ-sidecar pattern, but qwen35 uses the
@@ -915,8 +869,8 @@ fn load_weight_tensor(hfq: &HfqFile, gpu: &Gpu, name: &str, m: usize, k: usize) 
         // the weight bytes have already been uploaded to GPU (owned by
         // `wt.buf`) so the borrow no longer matters.
         if wt.gpu_dtype.supports_awq_sidecar() {
-            wt.awq_scale = load_awq_scale_for(hfq, gpu, &full_name, k)
-                .or_else(|| load_awq_scale_for(hfq, gpu, name, k));
+            wt.awq_scale = load_awq_scale(hfq, gpu, &full_name, k)
+                .or_else(|| load_awq_scale(hfq, gpu, name, k));
         }
         return Ok(wt);
     }
@@ -927,8 +881,8 @@ fn load_weight_tensor(hfq: &HfqFile, gpu: &Gpu, name: &str, m: usize, k: usize) 
             .unwrap_or_else(|| panic!("tensor not found: {name} or {full_name}"));
         let mut wt = load_weight_tensor_raw(gpu, info.quant_type, data, m, k)?;
         if wt.gpu_dtype.supports_awq_sidecar() {
-            wt.awq_scale = load_awq_scale_for(hfq, gpu, &full_name, k)
-                .or_else(|| load_awq_scale_for(hfq, gpu, name, k));
+            wt.awq_scale = load_awq_scale(hfq, gpu, &full_name, k)
+                .or_else(|| load_awq_scale(hfq, gpu, name, k));
         }
         Ok(wt)
     }
@@ -1445,13 +1399,13 @@ pub fn load_weights(hfq: &mut HfqFile, config: &Qwen35Config, gpu: &mut Gpu) -> 
     // attaching a sidecar here would have driven the 0.67 → 13.5 KLD
     // corruption documented at `docs/plans/awq_fix_claude.md` because
     // the spec-verify path used the non-AWQ `rotate_x_mq_batched`.
-    // Try each plausible tensor name; `load_awq_scale_for` returns
+    // Try each plausible tensor name; `load_awq_scale` returns
     // None when no sidecar exists, so this is a no-op for current
     // pre-CUDA-pipeline files.
     if output.gpu_dtype.supports_awq_sidecar() {
-        output.awq_scale = load_awq_scale_for(hfq, gpu, "lm_head.weight", config.dim)
-            .or_else(|| load_awq_scale_for(hfq, gpu, "model.language_model.lm_head.weight", config.dim))
-            .or_else(|| load_awq_scale_for(hfq, gpu, "model.language_model.embed_tokens.weight", config.dim));
+        output.awq_scale = load_awq_scale(hfq, gpu, "lm_head.weight", config.dim)
+            .or_else(|| load_awq_scale(hfq, gpu, "model.language_model.lm_head.weight", config.dim))
+            .or_else(|| load_awq_scale(hfq, gpu, "model.language_model.embed_tokens.weight", config.dim));
         eprintln!("  lm_head AWQ sidecar: {}",
             if output.awq_scale.is_some() { "attached" } else { "absent (no-op)" });
     }
@@ -1694,9 +1648,9 @@ fn load_output_into(
     // (spec-verify) route through AWQ-aware rotations on
     // `output.awq_scale.is_some()`. No-op on current files.
     if output.gpu_dtype.supports_awq_sidecar() {
-        output.awq_scale = load_awq_scale_for(hfq, gpu, "lm_head.weight", config.dim)
-            .or_else(|| load_awq_scale_for(hfq, gpu, "model.language_model.lm_head.weight", config.dim))
-            .or_else(|| load_awq_scale_for(hfq, gpu, "model.language_model.embed_tokens.weight", config.dim));
+        output.awq_scale = load_awq_scale(hfq, gpu, "lm_head.weight", config.dim)
+            .or_else(|| load_awq_scale(hfq, gpu, "model.language_model.lm_head.weight", config.dim))
+            .or_else(|| load_awq_scale(hfq, gpu, "model.language_model.embed_tokens.weight", config.dim));
         eprintln!("  lm_head AWQ sidecar: {}",
             if output.awq_scale.is_some() { "attached" } else { "absent (no-op)" });
     }
